+++ lib/active_record/connection_adapters/sqlserver/core_ext/calculations.rb
@@ -8,25 +8,6 @@ module ActiveRecord
     module SQLServer
       module CoreExt
         module Calculations
-          # Same as original except we don't perform PostgreSQL hack that removes ordering.
-          def calculate(operation, column_name)
-            return super unless klass.connection.adapter_name == "SQLServer"
-
-            if has_include?(column_name)
-              relation = apply_join_dependency
-
-              if operation.to_s.downcase == "count"
-                unless distinct_value || distinct_select?(column_name || select_for_count)
-                  relation.distinct!
-                  relation.select_values = [klass.primary_key || table[Arel.star]]
-                end
-              end
-
-              relation.calculate(operation, column_name)
-            else
-              perform_calculation(operation, column_name)
-            end
-          end
 
           private